Raise error and pass it to error handler
When we can not send all bytes at once. Basically, UDP sockets try to
send all bytes at once and if the message size is larger than that
atomic send operation it returns error (EMSGSIZE on linux) so we might
not have to care about this... But I'm just not sure about this.